Flask/requirements.txt
paramiko
python-logstash
mariadb